+++ R/gsgompertz.R
@@ -18,68 +18,33 @@ gsgCurve <- function(n, w){
   
   t <- 1:n
   # Cumulative
-  At <- w[4] * ((1 - exp(-w[2] * t)) * (1 + w[1] * exp(-w[2] * t))^-w[3])
+  # At <- w[4] * ((1 - exp(-w[2] * t)) * (1 + w[1] * exp(-w[2] * t))^-w[3])
+  At <- w[1] * ((1 - exp(-w[3] * t)) * (1 + w[2] * exp(-w[3] * t))^-w[4])
   # Adoption
   at <- diff(c(0, At))
-  Y <- cbind(At, at)
-  colnames(Y) <- c("Cumulative Adoption", "Adoption")
+  x <- cbind(At, at)
+  colnames(x) <- c("Cumulative Adoption", "Adoption")
   
-  return(Y)
+  return(x)
 }
 
-gsgInit <- function(x, loss, optim){
+gsgInit <- function(y, loss, optim, optsol, initpar, mscal){
   # Internal function: get initial values
   # We use Bass model paramters assuming c = 1 (see Bemmaor 1994)
-  # x in adoption per period
+  # y in adoption per period
   
   # calling bass estimates
-  what <- diffusionEstim(x, loss, pvalreps = 0, type = "bass", optim = optim)$w
+  what <- diffusionEstim(y, loss, pvalreps = 0, type = "bass", optim = optim,
+                        optsol = optsol, initopt = initpar, mscal = mscal)$w
 
   # Bemmaor shows that if a = 1, Beta = p/q and b = p + q
-  a <- what[1] / what[2] # the shape parameter beta
-  b <- what[1] + what[2] # the scale parameter b
-  m <- what[3] # the market size
+  a <- what[2] / what[3] # the shape parameter beta
+  b <- what[2] + what[3] # the scale parameter b
+  m <- what[1] # the market size
   c <- 1 # this is the shifting parameter alpha
   
-  w <- c(a, b, c, m)
-  names(w) <- c("a", "b", "c", "m")
+  w <- c(m, a, b, c)
+  names(w) <- c("m", "a", "b", "c")
   
   return(w)
-}
-
-gsgCost <- function(w, x, loss, w.idx = rep(TRUE, 4), prew = NULL,
-                          cumulative = c(TRUE, FALSE)) {
-  # Internal function: cost function for numerical optimisation
-  # w, current parameters
-  # x, adoption per period
-  # loss, the l-norm (1 is absolute errors, 2 is squared errors)
-  # w.idx, logical vector with three elements. Use FALSE to not estimate
-  # respective parameter
-  # prew, the w of the previous generation - this is used for sequential fitting
-  # cumulative, use cumulative adoption or not
-  
-  cumulative <- cumulative[1]
-  n <- length(x)
-  
-  # If some elements of w are not optimised, sort out vectors
-  w.all <- rep(0, 4)
-  w.all[w.idx] <- w
-  
-  # If sequential construct total parameters
-  if (is.null(prew)){
-    gsgpw <- w.all    
-  } else {
-    gsgpw <- w.all + prew
-  }
-  
-  fit <- gsgCurve(n, gsgpw)
-
-  se <- getse(x, fit, loss, cumulative) # auxiliary.R
-  
-  # Ensure positive coefficients
-  if (any(gsgpw <= 0)){
-    se <- 10e200
-  }
-  
-  return(se)
 }